feat: closer fact checking
Doesn't wire it up the the model, but otherwise it works as intended.
I've got a nice progress bar and the results are displayed in a table.
Sentences are split decently, and we filter out copyright stuff.
Oh, and we're able to query the API with claims now.
Before, we'd send the claim without sanitizing it, which lead to invalid queries.